requirements.txt
Django==5.1.4
django-cors-headers==4.6.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
dnspython==2.7.0
docx2pdf==0.1.8
et_xmlfile==2.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
mongoengine==0.29.1
numpy==2.2.0
openpyxl==3.1.5
pandas==2.2.3
openai==1.70.0
django-cors-headers==4.6.0
PyJWT==2.10.1